Added some blinking animiations on line clear
I also seem to have nuked my input handling, it feels much more janky now. Going to have to overhaul that

+++ include/game/piecequeue.h
@@ -13,7 +13,7 @@ typedef struct QueueNode {
 typedef struct PieceQueue {
     PieceQueueNode* head;
     PieceQueueNode* tail;
-    size_t          size;
+    i64             size;
 } PieceQueue;
 
 PieceQueue pq_create(void);
